+++ app/src/main/java/com/localmusic/player/playlist/PlaylistStore.kt
@@ -1,0 +1,7 @@
+package com.localmusic.player.playlist
+
+interface PlaylistStore {
+    fun playlists(): List<M3uPlaylist>
+    fun save(playlist: M3uPlaylist): List<M3uPlaylist>
+    fun delete(name: String): List<M3uPlaylist>
+}